feat: move tests next to files
